# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Elkasir API
  version: 0.2.0
  description: >-
    Kontrak OpenAPI Elkasir — sumber kebenaran tipe yang dikonsumsi web app.
    Diturunkan langsung dari handler Go + DTO (json tags). Money fields adalah
    integer (int64). Auth memakai JWT Bearer (header Authorization: Bearer
    <accessToken>).
servers:
  - url: http://localhost:8081
    description: Dev lokal
security:
  - bearerAuth: []
...
paths:
...
  /staff:
    get:
      tags: [Staff]
      operationId: listStaff
      summary: Daftar staf
      responses:
        "200":
          description: Daftar staf terpaginasi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedStaff" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Staff]
      operationId: createStaff
      summary: Buat staf (owner/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffCreateInput" }
      responses:
        "201":
          description: Staf dibuat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /staff/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [Staff]
      operationId: getStaff
      summary: Detail staf
      responses:
        "200":
          description: Detail staf
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Staff]
      operationId: updateStaff
      summary: Perbarui staf (owner/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffUpdateInput" }
      responses:
        "200":
          description: Staf diperbarui
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Staff]
      operationId: deleteStaff
      summary: Hapus staf (owner/admin)
      responses:
        "204": { description: Staf dihapus }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /staff/{id}/reset-password:
    parameters:
      - $ref: "#/components/parameters/PathId"
    post:
      tags: [Staff]
      operationId: resetStaffPassword
      summary: Reset password staf (owner/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 6 }
      responses:
        "204": { description: Password direset }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /staff/{id}/pin:
    parameters:
      - $ref: "#/components/parameters/PathId"
    put:
      tags: [Staff]
      operationId: setStaffPin
      summary: Set/hapus PIN persetujuan supervisor (owner/admin; hanya untuk role supervisor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin: { type: string, description: "PIN 4–6 digit; string kosong menghapus PIN." }
      responses:
        "204": { description: PIN disimpan/dihapus }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─────────────────────────── Admin Users ───────────────────────────
...
  /transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: Daftar transaksi (dengan filter)
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - { name: status, in: query, schema: { type: string } }
        - { name: source, in: query, schema: { type: string, enum: [cashier, self_order] } }
        - { name: paymentMethod, in: query, schema: { type: string, enum: [cash, qris] } }
        - { name: search, in: query, schema: { type: string } }
        - { name: from, in: query, description: "RFC3339 atau YYYY-MM-DD", schema: { type: string } }
        - { name: to, in: query, description: "RFC3339 atau YYYY-MM-DD", schema: { type: string } }
      responses:
        "200":
          description: Daftar transaksi terpaginasi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedTransaction" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Buat transaksi kasir (staf POS, idempoten)
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Kunci idempotensi wajib; permintaan ulang dengan body sama mengembalikan hasil lama.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionCreateInput" }
      responses:
        "201":
          description: Transaksi baru dibuat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "200":
          description: Replay idempotency (hasil transaksi sebelumnya)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [Transactions]
      operationId: getTransaction
      summary: Detail transaksi
      responses:
        "200":
          description: Detail transaksi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─────────────────────────── Cash Movements ───────────────────────────
...
  /pos/approvals/verify-pin:
    post:
      tags: [Auth]
      operationId: verifySupervisorPin
      summary: Verifikasi PIN supervisor (staf POS, rate-limited) untuk approve-in-place
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties: { pin: { type: string, description: "PIN 4–6 digit" } }
      responses:
        "200":
          description: Identitas supervisor penyetuju
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupervisorApprovalRef" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { description: Terlalu banyak percobaan (rate limit) }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 20 }
    Offset:
      name: offset
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    BadRequest:
      description: Permintaan tidak valid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Tidak terautentikasi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Tidak berwenang
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Sumber daya tidak ditemukan
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Konflik (duplikasi / kondisi tidak terpenuhi)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unprocessable:
      description: Entitas tidak dapat diproses (mis. stok kurang)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - validation_error
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - unprocessable
                - rate_limited
                - internal
            message: { type: string }
            details: {}
    Pagination:
      type: object
      required: [total, limit, offset]
      properties:
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }

    # ── Auth ───────────────────────────────────────────────────
...
    Staff:
      type: object
      required: [id, name, username, email, role, status, hasPin, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        username: { type: string }
        email: { type: string }
        role: { type: string, enum: [cashier, supervisor] }
        status: { type: string, enum: [active, inactive] }
        hasPin: { type: boolean, description: "Supervisor approval PIN is set (admin can spot who still needs one)." }
        createdAt: { type: string, format: date-time }
    StaffCreateInput:
      type: object
      required: [name, username, password, role]
      properties:
        name: { type: string }
        username: { type: string }
        email: { type: string }
        password: { type: string, minLength: 6 }
        role: { type: string, enum: [cashier, supervisor] }
        status: { type: string, enum: [active, inactive] }
    StaffUpdateInput:
      type: object
      required: [name, username, role]
      properties:
        name: { type: string }
        username: { type: string }
        email: { type: string }
        role: { type: string, enum: [cashier, supervisor] }
        status: { type: string, enum: [active, inactive] }

    # ── AdminUser ──────────────────────────────────────────────
...
    TransactionItem:
      type: object
      required: [productName, category, price, quantity, lineTotal]
      properties:
        productId: { type: string }
        productName: { type: string }
        category: { type: string }
        price: { type: integer, format: int64 }
        quantity: { type: integer, format: int32 }
        lineTotal: { type: integer, format: int64 }
        note: { type: string }
    Transaction:
      type: object
      required:
        [id, code, orderType, source, paymentMethod, status, subtotal, discount,
         tax, total, amountReceived, changeAmount, createdAt, items]
      properties:
        id: { type: string }
        code: { type: string }
        shiftId: { type: string }
        tableId: { type: string }
        selfOrderId: { type: string }
        cashierId: { type: string }
        orderType: { type: string, enum: [dineIn, takeaway] }
        source: { type: string, enum: [cashier, self_order] }
        paymentMethod: { type: string, enum: [cash, qris] }
        status: { type: string }
        subtotal: { type: integer, format: int64 }
        discount: { type: integer, format: int64 }
        tax: { type: integer, format: int64, description: "PPN" }
        serviceCharge: { type: integer, format: int64, description: "Biaya layanan 2% (rounded)." }
        gatewayFee: { type: integer, format: int64, description: "Biaya gateway QRIS (0 utk cash/kasir)." }
        serviceLine: { type: integer, format: int64, description: "Layanan = serviceCharge + gatewayFee." }
        total: { type: integer, format: int64 }
        amountReceived: { type: integer, format: int64 }
        changeAmount: { type: integer, format: int64 }
        customerNote: { type: string }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/TransactionItem" }
    TransactionCreateItem:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, format: int32 }
        note: { type: string }
    TransactionCreateInput:
      type: object
      required: [items, paymentMethod]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TransactionCreateItem" }
        discount: { type: integer, format: int64 }
        paymentMethod: { type: string, enum: [cash, qris] }
        amountReceived: { type: integer, format: int64 }
        tableId: { type: string }
        orderType: { type: string, enum: [dineIn, takeaway] }
        discountApprovedBy: { type: string }
        customerNote: { type: string }

    # ── CashMovement ───────────────────────────────────────────
...
    SupervisorApprovalRef:
      type: object
      required: [approvedById, approvedByName]
      properties:
        approvedById: { type: string }
        approvedByName: { type: string }
...
    PaginatedStaff:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Staff" }
        pagination: { $ref: "#/components/schemas/Pagination" }
...
    PaginatedTransaction:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Transaction" }
        pagination: { $ref: "#/components/schemas/Pagination" }
...